requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
huggingface_hub
ipykernel
chromadb
python-dotenv
docarray
tiktoken
openai
llama-index
llama-index-core
weaviate-client==4.5.7
streamlit==1.30.0
youtube_transcript_api
pathlib
pytube